requirements.txt
Flask
Flask-SocketIO
Flask-Cors
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
Pillow
bcrypt
